+++ data/services/ambari/0.2.2.0/rewrite.xml
@@ -1,0 +1,37 @@
+<!--
+   Licensed to the Apache Software Foundation (ASF) under one or more
+   contributor license agreements.  See the NOTICE file distributed with
+   this work for additional information regarding copyright ownership.
+   The ASF licenses this file to You under the Apache License, Version 2.0
+   (the "License"); you may not use this file except in compliance with
+   the License.  You may obtain a copy of the License at
+       http://www.apache.org/licenses/LICENSE-2.0
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+-->
+<rules>
+    <rule dir="IN" name="AMBARI/ambari/inbound" pattern="*://*:*/**/ambari/api/v1/{**}?{**}">
+        <rewrite template="{$serviceUrl[AMBARI]}/api/v1/{**}?{**}"/>
+    </rule>
+    <rule dir="OUT" name="AMBARI/ambari/href/outbound">
+        <match pattern="*://*:*/api/{**}?{**}"/>
+        <rewrite template="{$frontend[url]}/ambari/api/{**}?{**}"/>
+    </rule>
+    <rule dir="OUT" name="AMBARI/ambari/context_path/outbound">
+        <match pattern="/{**}"/>
+        <rewrite template="{$frontend[path]}/ambari/{**}"/>
+    </rule>
+    <filter name="AMBARI/ambari/api/outbound">
+        <content type="text/plain" asType="application/json">
+            <apply path="$.**.href" rule="AMBARI/ambari/href/outbound"/>
+            <apply path="$.**.context_path" rule="AMBARI/ambari/context_path/outbound"/>
+        </content>
+    </filter>
+    <filter name="AMBARI/ambari/api/inbound">
+        <content type="application/x-www-form-urlencoded" asType="application/octet-stream">
+        </content>
+    </filter>
+</rules>
