+++ dp_cluster_reg/base/cluster.py
@@ -1,0 +1,4 @@
+
+class BaseCluster(object):
+    def has_service(self, service_name):
+        return service_name in self.service_names()
