+++ dp_cluster_reg/exceptions.py
@@ -1,0 +1,10 @@
+class UnexpectedHttpCode(Exception):
+    pass
+
+
+class NoClusterFound(Exception):
+    pass
+
+
+class NoConfigFound(Exception):
+    pass
